A wrapper function for one-step forecasting with a static model (no refitting)

This function faciliates the forecasting approach in Ward et al. 2014. It has two operating modes. If num_ahead is set, then it forecasts the last num_ahead points of the time series:

fit the model to all but the last num_ahead time points

make 1-step ahead forecasts for the num_ahead time points

forecast_one_step_static(fun, timeseries, num_ahead = 5, ...)

Arguments

fun
a function for doing the forecasting. It should have arguments:

training

the data for training the model

observed

the observed values to be forecasted

...

any optional arguments

and should return a data.frame with at least the observed and predicted values (other columns are optional, and may be specific to the forecasting method)

timeseries
the time series to forecast

num_ahead
the number of points at the end of the time series to forecast

...
arguments to pass to fun

Value

a data.frame of the observed and predicted values, with any other remaining values returned by fun. If and error occurs, then NA values for the observed and predicted