Make forecasts using gaussian process regression

gausspr_ts fits a timeseries model using gausspr

gausspr_ts(timeseries, num_ahead = 5, level = 95, frequency = 1,
  kernel = "rbfdot", kpar = "automatic")

Arguments

timeseries
the time series to forecast

num_ahead
the number of points at the end of the time series to forecast

level
Confidence level for prediction intervals.

frequency
the number of observations per unit of time.

kernel
the kernel function used in training and predicting. This parameter can be set to any function, of class kernel, which computes a dot product between two vector arguments. kernlab provides the most popular kernel functions which can be used by setting the kernel parameter to the following strings:

rbfdot Radial Basis kernel function "Gaussian"

polydot Polynomial kernel function

vanilladot Linear kernel function

tanhdot Hyperbolic tangent kernel function

laplacedot Laplacian kernel function

besseldot Bessel kernel function

anovadot ANOVA RBF kernel function

splinedot Spline kernel

The kernel parameter can also be set to a user defined function of class kernel by passing the function name as an argument.

kpar
the list of hyper-parameters (kernel parameters). This is a list which contains the parameters to be used with the kernel function. Valid parameters for existing kernels are :

sigma inverse kernel width for the Radial Basis kernel function "rbfdot" and the Laplacian kernel "laplacedot".

degree, scale, offset for the Polynomial kernel "polydot"

scale, offset for the Hyperbolic tangent kernel function "tanhdot"

sigma, order, degree for the Bessel kernel "besseldot".

sigma, degree for the ANOVA kernel "anovadot".

Hyper-parameters for user defined kernels can be passed through the kpar parameter as well.

Value

a data.frame of the mean forecasts, the observed values, and the lower and upper CI levels (if an error occurs, then just NA values)